Add export/import of switchTab function
