+++ index.html
@@ -9,7 +9,7 @@
 <body>
   <div class="content">
     <nav class="navbar">
-      <a class="navbar__brand" href="./">
+      <a class="navbar__brand" href="./index.html">
         <img class="brand__logo" src="https://img.icons8.com/ios/32/000000/coffee-to-go.png"/>
         <img class="brand__logo" src="https://img.icons8.com/ios/50/000000/bread-loaf.png"/>
         <img class="brand__logo" src="https://img.icons8.com/ios/50/000000/sunny-side-up-eggs.png"/>
@@ -49,6 +49,7 @@
     </div>
   </div>
 
- <script src="main.js"></script>
+ <script src="./main.js"></script>
+ <!-- <script src="./main.js"></script> -->
 </body>
 </html>
